int to int64

+++ metaserver2/src/Routing.hs
@@ -5,24 +5,25 @@ module Routing where
 
 import Models
 import Servant
+import Data.Int (Int64)
 
 type BookEP = "books" :> Get '[JSON] [Book]
   :<|> "books" :> ReqBody '[JSON] Book :> Post '[JSON] Book
-  :<|> "books" :> Capture "id" Int :> Get '[JSON] Book
-  :<|> "books" :> Capture "id" Int :> ReqBody '[JSON] Book :> Put '[JSON] ()
-  :<|> "books" :> Capture "id" Int :> Delete '[JSON] ()
+  :<|> "books" :> Capture "id" Int64 :> Get '[JSON] Book
+  :<|> "books" :> Capture "id" Int64 :> ReqBody '[JSON] Book :> Put '[JSON] ()
+  :<|> "books" :> Capture "id" Int64 :> Delete '[JSON] ()
 
 type CircleEP = "circles" :> Get '[JSON] [Circle]
   :<|> "circles" :> ReqBody '[JSON] Circle :> Post '[JSON] Circle
-  :<|> "circles" :> Capture "id" Int :> Get '[JSON] Circle
-  :<|> "circles" :> Capture "id" Int :> ReqBody '[JSON] Circle :> Put '[JSON] ()
-  :<|> "circles" :> Capture "id" Int :> Delete '[JSON] ()
+  :<|> "circles" :> Capture "id" Int64 :> Get '[JSON] Circle
+  :<|> "circles" :> Capture "id" Int64 :> ReqBody '[JSON] Circle :> Put '[JSON] ()
+  :<|> "circles" :> Capture "id" Int64 :> Delete '[JSON] ()
 
 type WriterEP = "writers" :> Get '[JSON] [Writer]
   :<|> "writers" :> ReqBody '[JSON] Writer :> Post '[JSON] Writer
-  :<|> "writers" :> Capture "id" Int :> Get '[JSON] Writer
-  :<|> "writers" :> Capture "id" Int :> ReqBody '[JSON] Writer :> Put '[JSON] ()
-  :<|> "writers" :> Capture "id" Int :> Delete '[JSON] ()
+  :<|> "writers" :> Capture "id" Int64 :> Get '[JSON] Writer
+  :<|> "writers" :> Capture "id" Int64 :> ReqBody '[JSON] Writer :> Put '[JSON] ()
+  :<|> "writers" :> Capture "id" Int64 :> Delete '[JSON] ()
 
 type TagEP = "tags" :> Get '[JSON] [String]
